requirements.txt
pandas==2.2.2
streamlit==1.35.0
numpy==1.26.4
anthropic==0.25.0
python-dotenv==1.0.1
